requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.7
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==7.1.1
forex-python==1.8
idna==3.10
inflection==0.5.1
mccabe==0.7.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
stripe==11.3.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
